Add system info and the ability to write modified configuration.
We can now save configuration (good) but it needs to save the whole section, not the options which are only
changed from the defaults. This means that there's a large number of things that are duplicated and this
might lead to later having some confusion with what is set or not set from the default.

Considering adding a "reset to defaults" button for each section that simply removes the section that has
come under the knife. There's a few good reasons to do this.

in runserber.py we need to declare both the local and base config so that writing works.