sqash: 3 commits
LibServer: use a custom config class in Flask:

We're going to need a custom config class. This lets us use the configuration from disk
while still making sure that we have a good amount of configurability. I need to make a
way to load configuration (or reload configuration) from disk when there is a change to
the configuration through the admin panel (or we declare that we must have a restart to
see those changes applied). I'm considering making the admin panel touch this directly
and load them from there (see get_config_parser in admin.py)

Layout: Test using the changed configuration

Storage: use the namespaced options.

Storage should be namespaced so that We can later use autofs (which will be how we load
and unload filesystems at will) such that we can have loadable collections.