Re-lay out the tree
This is something I should have done a long time ago. This makes the installation sequence so much easier (now I just need to copy some files into the right place and there isn't a whole big bit of brou-ha-ha that comes along with that).

Let-s consider the fact that before this, installing was a pain in the ass and involved a lot of hand-waving. Now, you check out the repo and you're golden.

This is what I wanted it to be in the first place. Fantastic.