admin: Let there be EXTRA view!
There is now an untemplated view of any section that can be stumbled upon in the configuration. This is the same view that is presented if there is no specific template for the section, and is going to be used in the interim for enabling the internal storage on the pi once I get that working.