Make the login process not fall over when you botch logging in the first time.
The reason was that there was only a check to see if it was in the get parameters. Instead, after a little logic rework, the thing that I was going to do is now actually there.

We take the next parameter and if we've seen it before we just hand it off to the form and haul it along with the post data. Either one is fine, and the get argument overrides the POST argument.