Apple's Captive Network Detection Sucks.
So this is a really bad way of handling this. It's the best way I can think of though.

We consume a wildcard DNS that attempts to siphon all the content into the librarybox httpd. If it sees the UA CaptiveNetworkSupport, it will turn around and shovel it off to a special blackhole of 200 OK.